requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
httpx==0.27.2
anthropic==0.39.0
google-genai==0.3.0
replicate==1.0.3
pydantic==2.9.2
python-dotenv==1.0.1
